requirements.txt
opencv-python
opencv-contrib-python
numpy
pyttsx3
requests
Pillow
